+++ src/main.tsx
@@ -3,13 +3,16 @@ import ReactDOM from 'react-dom/client'
 import { ChakraProvider } from '@chakra-ui/react'
 import App from './App'
 import './index.css'
+import { LiffProvider } from './hooks/LiffProvider'
 import LiffCheck from './LiffCheck'
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
     <ChakraProvider>
-      <App />
-      <LiffCheck />
+      <LiffProvider mock liffId="aaa">
+        <App />
+        <LiffCheck />
+      </LiffProvider>
     </ChakraProvider>
   </React.StrictMode>
 )
